+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Paratrooper Commander
   - Rapid Commander
 - Loadout name to GWO panel
+- Player information to GWO panel
 
 ### Changed